Refactor responses
