+++ lib/chainweb/pact/spv_response.ex
@@ -1,0 +1,13 @@
+defmodule Kadena.Chainweb.Pact.SPVResponse do
+  @moduledoc """
+  `SPVResponse` struct definition.
+  """
+
+  @type response :: String.t()
+
+  @type t :: %__MODULE__{response: response()}
+
+  defstruct [:response]
+
+  def new(response) when is_binary(response), do: struct(%__MODULE__{}, response: response)
+end
