--- lib/chainweb/resource.ex
@@ -1,10 +1,0 @@
-defmodule Kadena.Chainweb.Resource do
-  @moduledoc """
-  Specifies contracts to build Chainweb resources.
-  """
-
-  @type attrs :: map()
-  @type resource :: struct()
-
-  @callback new(attrs()) :: resource()
-end
