+++ test/support/fixtures/chainweb/send.json
@@ -1,0 +1,6 @@
+{
+  "requestKeys": [
+    "ATGCYPMNzdGcFh9Iik73KfMkgURIxaF91Ze4sHFsH8Q",
+    "JHgnKe5Wd4hNIb7a6bIhm4ifxsYFzVGtAMyi_TEO-oM"
+  ]
+}
